pyproject.toml
[project]
name = "openclow"
version = "0.1.0"
description = "AI-Powered Development Orchestration Platform"
requires-python = ">=3.12"
dependencies = [
    # AI Agent
    "claude-agent-sdk>=0.1.50",
    # Chat providers
    "aiogram>=3.4",
    "slack-bolt>=1.18",
    "slack-sdk>=3.27",
    # Task Queue
    "arq>=0.26",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "alembic>=1.13",
    # API
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    # Web chat
    "assistant-stream>=0.0.31",
    "python-jose[cryptography]>=3.3",
    "bcrypt>=4.0",
    # Settings
    "pydantic-settings>=2.5",
    # MCP
    "mcp[cli]>=1.20",
    "mcp-server-git>=2025.1",
    # Logging
    "structlog>=24.0",
    # Utilities
    "python-slugify>=8.0",
    "httpx>=0.27",
    "redis>=5.0",
    "pyjwt>=2.8",  # GitHub App installation-token JWT mint (per-chat-instances research.md §3)
    # Voice transcription
    "faster-whisper>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "watchfiles>=0.24",
]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
